Move modified_widgets to widgets
