+++ scwx-qt/source/scwx/qt/ui/widgets/focused_double_spin_box.hpp
